+++ web-frontend/modules/database/components/view/form/FormViewFieldLinkRow.vue
@@ -3,6 +3,7 @@
     <PaginatedDropdown
       :fetch-page="fetchPage"
       :value="dropdownValue"
+      :initial-display-name="initialDisplayName"
       :class="{ 'dropdown--error': touched && !valid }"
       :fetch-on-open="lazyLoad"
       @input="updateValue($event)"
@@ -44,6 +45,9 @@ export default {
   computed: {
     dropdownValue() {
       return this.value.length === 0 ? null : this.value[0].id
+    },
+    initialDisplayName() {
+      return this.value.length === 0 ? null : this.value[0].value
     },
   },
   methods: {